backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secretary-partner-ai"
version = "0.1.0"
description = "Brain Dump Partner - ADHD向け自律型秘書AI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Secretary Partner AI Team" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",

    # Agent Framework
    "google-adk>=1.2.0",

    # LLM
    "litellm>=1.55.0",
    "boto3>=1.35.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",

    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # GCP (optional for production)
    "google-cloud-firestore>=2.20.0",
    "google-cloud-speech>=2.28.0",
    "firebase-admin>=6.6.0",

    # Scheduler
    "apscheduler>=3.11.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

local = [
    # For local speech-to-text (optional)
    "openai-whisper>=20240930",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: End-to-end tests with real API calls",
    "integration: Integration tests with real database",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
ignore_errors = true
